fix: mitigate geofence flapping (#6)
Because lat and long must be processed for geofence changes
individually, it's possible that when crossing geofences, there may be
some flapping. In the scenario where a user does not manually define a
cooldown or utilize garage status checks, this can cause the app to spam
the garage with commands during a brief window.

This adds a base cooldown to allow a full geofence crossing to take
place after the first garage command is sent to mitigate the flapping.