feat: enable any mqtt location trigger (#12)
Add support for any tracker that can publish to an MQTT broker, such as
TeslaMate or OwnTracks. This will enable usage for any vehicle when
using a tracking app on a phone to publish location data to an MQTT
broker.

This includes breaking changes to the config file, and users should
always check the examples directory for how to structure config files.