feat: allow duplicate trackers (#37)
### Summary

Allow duplicate trackers

### Issue
[Issue 36](https://github.com/brchri/tesla-geogdo/issues/36)

### Context

There may be scenarios where a user wishes to control multiple doors
with a single tracker (vehicle), but currently when a topic message is
consumed, the code looks for the first matching tracker and then breaks,
so any other doors with the same tracker value would never execute. This
removes the breaks to search through all trackers for a match.